requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
fpdf==1.7.2
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-groq==0.3.6
langchain-text-splitters==0.3.9
langsmith==0.4.9
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
numpy==2.2.6
orjson==3.11.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
stack-data==0.6.3
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
yarl==1.20.1
zstandard==0.23.0
